% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_across_terms.R
\name{combine_across_terms}
\alias{combine_across_terms}
\title{Combine a particular data set across multiple terms}
\usage{
combine_across_terms(data_structure, data_folder)
}
\arguments{
\item{data_structure}{string}

\item{data_folder}{string}
}
\description{
Combine a particular data set across multiple terms
}
