% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{Extract comments from an SETC report}
\usage{
get_comments(report_text, course_id = "")
}
\arguments{
\item{report_text}{a vector of strings, usually
generated by \code{get_report_text}}

\item{course_id}{a string that will be included as the course_id}
}
\value{
A tibble in which each row corresponds to a single
line of commentary, and the variables include the comment text
and the provided \code{course_id}.
}
\description{
\code{get_comments} extracts the comments from the text of
an SETC report.
}
\details{
\code{get_comments} will look in \code{report_text} for the text
"Do you have any further comments?"  It will then treat (almost)
every line after that as a comment.

If \code{report_text} does not include the text
"Do you have any further comments?", \code{get_comments}
will return NULL.
}
