% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instructor_data.R
\name{get_instructor_data}
\alias{get_instructor_data}
\title{Read in additional instructor information as a tibble}
\usage{
get_instructor_data(instructor_file = NULL)
}
\arguments{
\item{instructor_file}{An optional string giving the location of the
CSV file with additional instructor information.}
}
\value{
If instructor_file is provided, a tibble (data frame)
containing instructor information.

If \code{instructor_file}
is not provided, \code{get_instructor_data} returns NULL.
}
\description{
\code{get_instructor_data} reads additional instructor information
from \code{instructor_file}, if it is provided.
}
\details{
If provided, \code{instructor_file} should be a CSV file in which the top
row gives variable names, and each additional row corresponds
to an instructor. Variables should include \code{instructor_id}
so that this data can be linked to other SETC data. To avoid name
conflicts, it is advisable for all variables to have the prefix
\code{instructor_}.
}
\examples{
get_instructor_data()
\dontrun{
get_instructor_data("..\\data\\master\\master intructor list.csv")
}
}
