% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instructors.R
\name{get_instructors}
\alias{get_instructors}
\title{Read in the master instructor list as a tibble}
\usage{
get_instructors(instructor_file, output_file = NULL)
}
\arguments{
\item{instructor_file}{filename}

\item{output_file}{filename}
}
\description{
Read in the master instructor list as a tibble
}
