% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_questions.R
\name{get_questions}
\alias{get_questions}
\title{Title}
\usage{
get_questions(question_file, output_file = NULL)
}
\arguments{
\item{question_file}{a file name}

\item{output_file}{a file name}
}
\description{
Title
}
