% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response_rate.R
\name{get_response_rate}
\alias{get_response_rate}
\title{Finds response rate information in SETC report}
\usage{
get_response_rate(report_text, row_name = "Responded")
}
\arguments{
\item{report_text}{The text of an SETC report, normally
obtained with \code{\link{get_report_text}}.}

\item{row_name}{A string, either "Responded" or "Invited"}
}
\value{
A number, either the number of students invited
  to the SETC survey, or the number who responded.
}
\description{
Finds response rate information in SETC report
}
