% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_distribution.R
\name{get_score_distribution}
\alias{get_score_distribution}
\title{Find and extract score distribution from SETC report}
\usage{
get_score_distribution(report_text, question)
}
\arguments{
\item{report_text}{string}

\item{question}{tibble}
}
\description{
Find and extract score distribution from SETC report
}
