% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_mean.R
\name{get_score_mean}
\alias{get_score_mean}
\title{Extract mean score from SETC report text}
\usage{
get_score_mean(report_text, question)
}
\arguments{
\item{report_text}{string}

\item{question}{tibble}
}
\description{
Extract mean score from SETC report text
}
