% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_term.R
\name{get_term}
\alias{get_term}
\title{Title}
\usage{
get_term(input, from = "report_name")
}
\arguments{
\item{input}{A string.}

\item{from}{A string describing the nature of the input. Currently
supported values include "report_name"}
}
\value{
A string identifying a term, e.g. "Fa18"
}
\description{
Title
}
