% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_course_id.R
\name{is_course_id}
\alias{is_course_id}
\title{Check whether input is a valid course_id}
\usage{
is_course_id(course_id, strict = FALSE)
}
\arguments{
\item{course_id}{string}

\item{strict}{logical}
}
\value{
logical
}
\description{
Check whether input is a valid course_id
}
\examples{
is_course_id("ECON381D1Fa19")
}
