% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_instructor_data.R
\name{is_instructor_data}
\alias{is_instructor_data}
\title{Check for valid instructor data}
\usage{
is_instructor_data(dat)
}
\arguments{
\item{dat}{A tibble (data frame) that supposedly contains
valid instructor data}
}
\value{
\code{TRUE} if \code{dat} is a valid instructor data set,
  \code{FALSE} if not.
}
\description{
\code{is_instructor_data} checks its input to see if it is
a tibble (or data frame) in which each row contains a
unique value of \code{dat$instructor_id}.
}
