% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_instructor_id.R
\name{is_instructor_id}
\alias{is_instructor_id}
\title{Checks for a valid instructor_id (to be added)}
\usage{
is_instructor_id(instructor_id, strict = FALSE)
}
\arguments{
\item{instructor_id}{string}

\item{strict}{logical}
}
\description{
Checks for a valid instructor_id (to be added)
}
