% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_question_data.R
\name{is_question_data}
\alias{is_question_data}
\title{Check for valid question data}
\usage{
is_question_data(dat)
}
\arguments{
\item{dat}{A tibble (data frame) that supposedly contains
valid question data}
}
\value{
\code{TRUE} if \code{dat} is a valid question data set,
  \code{FALSE} if not.
}
\description{
\code{is_question_data} checks its input to see if it is
a tibble (or data frame) in which each row contains a
unique value of \code{dat$question_id}.
}
