% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_setc_data.R
\name{pool_setc_data}
\alias{pool_setc_data}
\title{Pool multiple terms of SETC data into a single data set}
\usage{
pool_setc_data(data_structure, data_folder)
}
\arguments{
\item{data_structure}{A string describing an SETC data structure
Possible values include "course" or "comments".}

\item{data_folder}{A string, the folder in which the data can be found.}
}
\value{
A tibble (data frame) constructed by \code{rbind}ing
every CSV file in \code{data_folder} with a filename of the
form \code{data_structure_TERM.csv} where \code{TERM} is a valid term (e.g. "Fa16")
}
\description{
\code{pool_setc_data} combines multiple terms of SETC
data into a single pooled data set.
}
\details{
This function is designed to be called after
\code{\link{get_setc_data}}, which extracts SETC
data from the PDF files, and \code{\link{setc_save}},
which saves the data to CSV files. Once the term-by-term
CSV files have been generated,
use this function to pool them into a single data file.
}
\examples{
\dontrun{
pool_setc_data("course", "../data/")
}
}
\seealso{
\code{\link{get_setc_data}}, \code{\link{setc_save}}
}
