% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{Find and extract comments from the text of an SETC report}
\usage{
get_comments(report_text, course_id = "")
}
\arguments{
\item{report_text}{A vector of strings, usually
generated by \code{\link{get_report_text}}.}

\item{course_id}{An optional string that will be
treated as the course ID.}
}
\value{
A tibble (data frame) in which each row corresponds to a single
line of commentary, and the variables include:
  \describe{
  \item{\code{course_id}}{The course ID.}
  \item{\code{comment_numbers}}{Rows are numbered sequentially.}
  \item{\code{comment_text}}{The text of the comment.}
  \item{\code{instructor_id}, \code{term_id}, \code{course_number}}{Additional information about the course and instructor.}
  }
}
\description{
\code{get_comments} finds and extracts the student comments
from the text of an SETC report.
}
\details{
\code{get_comments} is normally called by \code{\link{get_setc_data}}
but can in principle be called on its own.

\code{get_comments} will look in \code{report_text} for the text
"Do you have any further comments?"  It will then treat (almost)
every line after that as a student comment.

If \code{report_text} does not include the text
"Do you have any further comments?", \code{get_comments}
will return NULL.
}
\seealso{
\code{\link{get_setc_data}}, \code{\link{get_report_text}}
}
