% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_id.R
\name{get_course_id}
\alias{get_course_id}
\title{Extract course_id from SETC report filename}
\usage{
get_course_id(report_name, term = NULL)
}
\arguments{
\item{report_name}{string}

\item{term}{string}
}
\value{
string
}
\description{
Extract course_id from SETC report filename
}
\examples{
get_course_id("BrianKrauth -ECON105D100-Fall2018.pdf", term="Fa18")
}
