% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_id.R
\name{get_course_id}
\alias{get_course_id}
\title{Extract a course ID from the SETC report filename.}
\usage{
get_course_id(report_name)
}
\arguments{
\item{report_name}{A string (vector) giving the filename(s) for
a SETC report}
}
\value{
A string (vector) giving the course ID that can be inferred
  from each \code{report_name}.
}
\description{
\code{get_course_id} infers a valid course ID from the file
name of a SETC report.
}
\details{
A course ID takes the form (for example) "ECON105D1Fa18"
where the components are the program code (ECON), course number
(105), section number (D1) and term (Fa18).

SETC filenames take on a standard format, (for example
"BrianKrauth -ECON105D100-Fall2018.pdf") from which the course
ID (ECON105D1Fa18) and instructor ID (BrianKrauth) can be
extracted.
}
\examples{
get_course_id("BrianKrauth -ECON105D100-Fall2018.pdf")
}
\seealso{
\code{\link{is_course_id}}, \code{\link{get_instructor_id}}.
}
