% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_number.R
\name{get_course_number}
\alias{get_course_number}
\title{Get course number}
\usage{
get_course_number(report_name = NULL, term = NULL,
  course_id = get_course_id(report_name, term = term))
}
\arguments{
\item{report_name}{string}

\item{term}{string}

\item{course_id}{string}
}
\description{
Get course number
}
