% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instructor_id.R
\name{get_instructor_id}
\alias{get_instructor_id}
\title{Extract instructor_id from SETC report filename}
\usage{
get_instructor_id(report_name, term = NULL)
}
\arguments{
\item{report_name}{string}

\item{term}{string}
}
\description{
Extract instructor_id from SETC report filename
}
