% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_text.R
\name{get_report_text}
\alias{get_report_text}
\title{Get report text}
\usage{
get_report_text(report_path)
}
\arguments{
\item{report_path}{A string giving the full path to
a PDF file containing a course-lvel SETC report}
}
\value{
A vector of strings, the text of the report.
}
\description{
Get report text
}
