% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response_rate.R
\name{get_response_rate}
\alias{get_response_rate}
\title{Finds response rate information in SETC report}
\usage{
get_response_rate(report_text, row_name = "Responded")
}
\arguments{
\item{report_text}{SETC report text}

\item{row_name}{string}
}
\description{
Finds response rate information in SETC report
}
