% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_distribution.R
\name{get_score_distribution}
\alias{get_score_distribution}
\title{Find and extract score distribution from SETC report}
\usage{
get_score_distribution(report_text, question)
}
\arguments{
\item{report_text}{A vector of strings, normally
generated by \code{\link{get_report_text}}.}

\item{question}{A tibble (data frame) of information on
an SETC question,
normally generated by a single row in \code{\link{get_question_data}}.}
}
\value{
A numeric vector of length 5 giving the distribution
  of responses.
}
\description{
Find and extract score distribution from SETC report
}
