% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_mean.R
\name{get_score_mean}
\alias{get_score_mean}
\title{Extract mean score from SETC report text}
\usage{
get_score_mean(report_text, question)
}
\arguments{
\item{report_text}{A vector of strings, normally
generated by \code{\link{get_report_text}}}

\item{question}{A tibble (data frame) of information on
an SETC question,
normally generated by a single row in \code{\link{get_question_data}}.}
}
\value{
A numeric scalar, the mean response to
  the question given in \code{question}.
}
\description{
Extract mean score from SETC report text
}
