% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setc_data.R
\name{get_setc_data}
\alias{get_setc_data}
\title{Title}
\usage{
get_setc_data(term, term_folder = NULL, output_folder = NULL,
  question_file = NULL, instructor_file = NULL)
}
\arguments{
\item{term}{A string, either "Sp", "Su", or "Fa"}

\item{term_folder}{The folder containing SETC reports}

\item{output_folder}{The folder in which you want output to be written}

\item{question_file}{Optional question file}

\item{instructor_file}{filename}
}
\description{
Title
}
