% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_term.R
\name{get_term}
\alias{get_term}
\title{Title}
\usage{
get_term(report_name = NULL, term_folder = NULL)
}
\arguments{
\item{report_name, term_folder}{A vector of strings}
}
\value{
A string identifying a term, e.g. "Fa18"
}
\description{
Title
}
\examples{
get_term(term_folder = "1171 - Spring 2017")
}
