% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_term_folder_name.R
\name{get_term_folder_name}
\alias{get_term_folder_name}
\title{Extract term folder name from the term}
\usage{
get_term_folder_name(term)
}
\arguments{
\item{term}{A string or vector of strings agiving the term (e.g. "Sp17")}
}
\value{
A string or vector of strings giving the associated term
  folder (e.g., "1171 - Spring 2017")
}
\description{
\code{get_term_folder_name} extracts a term folder name
(e.g., "1171 - Spring 2017") from the term (e.g., "Sp17")
}
\examples{
get_term_folder_name("Sp17")
}
