% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_term_id.R
\name{get_term_id}
\alias{get_term_id}
\title{Extract term_id (e.g. 1171) from either a term or a course_id}
\usage{
get_term_id(term = stringr::str_sub(course_id, -4), course_id = NULL)
}
\arguments{
\item{term}{numeric}

\item{course_id}{string}
}
\description{
Extract term_id (e.g. 1171) from either a term or a course_id
}
\examples{
get_term_id(term="Fa17")
get_term_id(course_id = "ECON381D1Fa17")
}
