% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_course_id.R
\name{is_course_id}
\alias{is_course_id}
\title{Check whether input is a valid course ID}
\usage{
is_course_id(course_id, strict = FALSE)
}
\arguments{
\item{course_id}{A vector of strings}

\item{strict}{Optional logical scalar.  If \code{strict = TRUE},
the function is case-sensitive and only allows known
course section types (currently C,D,E and G).  If
\code{strict = FALSE} (the default), the function is
case sensitive and allows any course section type.}
}
\value{
A logical vector equalling \code{TRUE} if the
  corresponding element in \code{course_id} is a valid
  course ID, and \code{FALSE} otherwise.
}
\description{
\code{is_course_id} checks whether its input is a valid course ID.
}
\examples{
is_course_id("ECON381D1Fa19")
is_course_id("ECON381D1Fa19", strict = TRUE)
is_course_id("NOT A VALID COURSE ID")
}
