% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_instructor_id.R
\name{is_instructor_id}
\alias{is_instructor_id}
\title{Check whether input is a valid instructor ID}
\usage{
is_instructor_id(instructor_id, strict = FALSE)
}
\arguments{
\item{instructor_id}{A vector of strings}

\item{strict}{A logical scalar (currently ignored).}
}
\value{
A logical vector, each element of which is TRUE
 if the corresponding element of \code{instructor_id} is
 a valid instructor ID.
}
\description{
\code{is_instructor_id} checks to see if its input
is a valid instructor ID.  Instructor IDs should
be alphabetic strings.
}
\examples{
is_instructor_id("JoeSmith")
}
