% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_setc_data.R
\name{pool_setc_data}
\alias{pool_setc_data}
\title{Combine a particular data set across multiple terms}
\usage{
pool_setc_data(data_structure, data_folder)
}
\arguments{
\item{data_structure}{A string describing an SETC data structure
Possible values include "course" or "comments"}

\item{data_folder}{The folder in which the data can be found}
}
\value{
A tibble (data frame) constructed by \code{rbind}ing
every CSV file in \code{data_folder} with a filename of the
form \code{data_structure_TERM.csv} where \code{TERM} is a valid term (e.g. "Fa16")
}
\description{
\code{pool_setc_data} combines multiple terms of a SETC
data set generated by \code{get_setc_data()} into a single
pooled data set.
}
