% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setc_save.R
\name{setc_save}
\alias{setc_save}
\title{Title}
\usage{
setc_save(dat, as, output_folder)
}
\arguments{
\item{dat}{A tibble or data frame}

\item{as}{A string indicating the type of data.}

\item{output_folder}{The folder into which the data should be saved}
}
\description{
Title
}
