% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setc_spread.R
\name{setc_spread}
\alias{setc_spread}
\title{Spread question_and_course level data out to course level data}
\usage{
setc_spread(dat, stat = "mean")
}
\arguments{
\item{dat}{A tibble with course-and-question level SETC data}

\item{stat}{A string indicating the statistic.  Valid values are
"n", "mean", "median", and "sd"}
}
\value{
A tibble
}
\description{
Spread question_and_course level data out to course level data
}
